Support running without s2s certs
This is a different approach to #4; in #4, the default was set to the
empty string, which is ambiguous. In addition, it was unclear what the
use-case would be, because I was being dense.

If the tool is exclusively used for c2s testing, s2s certs don't matter;
their lack should not prevent the tool from starting up. In addition,
s2s certs are not needed for all s2s cases either: for instance, testing
hosts with dialback support is still possible without s2s certs.